api/requirements.txt
Flask==3.0.3
transformers==4.41.2
python-docx==1.1.2
pdfplumber==0.1.2